Using cargo workspaces + Removed buildx dependency + Cleaned up (#6)
* updated repo to use cargo workspaces + removed buildx dependency from the build process + added spink build process t
o Makefile + using nethost instead of cni + cleaned up readmes / added tars to it

* adding changes from the latest PR
