+++ .gitignore
@@ -1,4 +1,3 @@
-bin/
-test/out/img.tar
 **/target/
 .vscode
+test/
