+++ Cargo.toml
@@ -1,39 +1,6 @@
-[package]
-name = "containerd-shim-wasm"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-containerd-shim-wasmtime-v1 = { git = "https://github.com/Mossaka/runwasi", branch = "jiazho/precompile" }
-wit-bindgen-wasmtime = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "2f46ce4cc072107153da0cefe15bdc69aa5b84d0" }
-cloudevents-sdk  = { default-features = false, git = "https://github.com/Mossaka/sdk-rust.git", branch = "master" }
-url = "2.2.2"
-rouille = "3.5.0"
-uuid = "0.8.2"
-anyhow = "1.0"
-chrono = "0.4.19"
-log = "0.4"
-wasmtime = "^0.34"
-wasmtime-wasi = "^0.34"
-serde_json = "1.0"
-containerd-shim = "0.3.0"
-spin-http-engine =  { git = "https://github.com/Mossaka/spin", branch = "jiazho/spin-test-with-custom-engine" }
-spin-loader =  { git = "https://github.com/Mossaka/spin", branch = "jiazho/spin-test-with-custom-engine" }
-tempfile = "3.3.0"
-tokio = { version = "1.11", features = [ "rt" ] }
-# see problem: https://github.com/sfackler/rust-openssl/issues/1021#issuecomment-605602642
-openssl = { version = "0.10", features = ["vendored"] }
-
-[[bin]]
-name = "containerd-shim-aspdotnet-v1"
-path = "src/containerd-shim-aspdotnet-v1/main.rs"
-
-[[bin]]
-name = "containerd-shim-cehostshim-v1"
-path = "src/containerd-shim-cehostshim-v1/main.rs"
-
-[[bin]]
-name = "containerd-shim-spin-v1"
-path = "src/containerd-shim-spin-v1/main.rs"
+[workspace]
+members = [
+    "containerd-shim-aspdotnet-v1",
+    "containerd-shim-cehostshim-v1",
+    "containerd-shim-spin-v1"
+]
