+++ containerd-shim-cehostshim-v1/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "containerd-shim-cehostshim-v1"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+chrono = "0.4.19"
+cloudevents-sdk  = { default-features = false, git = "https://github.com/Mossaka/sdk-rust.git", branch = "master" }
+containerd-shim = "0.3.0"
+containerd-shim-wasmtime-v1 = { git = "https://github.com/Mossaka/runwasi", branch = "jiazho/precompile" }
+log = "0.4"
+rouille = "3.5.0"
+uuid = "0.8.2"
+wasmtime = "^0.34"
+wasmtime-wasi = "^0.34"
+wit-bindgen-wasmtime = { git = "https://github.com/bytecodealliance/wit-bindgen", rev = "2f46ce4cc072107153da0cefe15bdc69aa5b84d0" }
+serde_json = "1.0"
