+++ containerd-shim-cehostshim-v1/src/types.wit
@@ -1,0 +1,7 @@
+// General purpose error.
+enum error {
+    success,
+    error,
+}
+
+type payload = list<u8>
