+++ containerd-shim-cehostshim-v1/src/wasi-ce.wit
@@ -1,0 +1,3 @@
+use * from types
+
+ce-handler: function(event: string) -> expected<string, error>
