--- src/containerd-shim-cehostshim-v1/types.wit
@@ -1,7 +1,0 @@
-// General purpose error.
-enum error {
-    success,
-    error,
-}
-
-type payload = list<u8>
