--- src/containerd-shim-cehostshim-v1/wasi-ce.wit
@@ -1,3 +1,0 @@
-use * from types
-
-ce-handler: function(event: string) -> expected<string, error>
